+++ DLPrototype/Views/Find/FindDashboard.swift
@@ -292,1142 +292,7 @@ extension FindDashboard {
             }
         }
     }
-    
-    struct Inspector: View, Identifiable {
-        public let id: UUID = UUID()
-        public var entity: NSManagedObject
-        
-        private let panelWidth: CGFloat = 400
-        private var job: Job? = nil
-        private var project: Project? = nil
-        private var record: LogRecord? = nil
-        private var company: Company? = nil
-        private var person: Person? = nil
-        private var note: Note? = nil
-        private var task: LogTask? = nil
 
-        @EnvironmentObject public var nav: Navigation
-
-        var body: some View {
-            VStack(alignment: .leading) {
-                HStack {
-                    FancySubTitle(text: "Inspector")
-                    Spacer()
-                    FancyButtonv2(
-                        text: "Close",
-                        action: {nav.session.search.inspectingEntity = nil},
-                        icon: "xmark",
-                        showLabel: false,
-                        size: .tiny,
-                        type: .clear
-                    )
-                }
-                Divider()
-                    .padding(.bottom, 10)
-
-                if let job = job {
-                    InspectingJob(item: job)
-                } else if let record = record {
-                    InspectingRecord(item: record)
-                } else if let project = project {
-                    InspectingProject(item: project)
-                } else if let company = company {
-                    InspectingCompany(item: company)
-                } else if let person = person {
-                    InspectingPerson(item: person)
-                } else if let note = note {
-                    InspectingNote(item: note)
-                } else if let task = task {
-                    InspectingTask(item: task)
-                } else {
-                    Text("Unable to inspect this item")
-                }
-
-                Spacer()
-            }
-            .padding()
-            .frame(maxWidth: panelWidth)
-        }
-        
-        init(entity: NSManagedObject) {
-            self.entity = entity
-            
-            switch self.entity {
-            case let en as Job: self.job = en
-            case let en as Project: self.project = en
-            case let en as LogRecord: self.record = en
-            case let en as Company: self.company = en
-            case let en as Person: self.person = en
-            case let en as Note: self.note = en
-            case let en as LogTask: self.task = en
-            default: print("[error] FindDashboard.Inspector Unknown entity type=\(self.entity)")
-            }
-        }
-        
-        struct InspectingJob: View {
-            public var item: Job
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Job")
-                        Spacer()
-                    }
-                    .help("Type: Job entity")
-                    Divider()
-
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "number").symbolRenderingMode(.hierarchical)
-                        Text(item.jid.string)
-                        Spacer()
-                    }
-                    .help("ID: \(item.jid.string)")
-                    Divider()
-                    
-                    if let date = item.created {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.lastUpdate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Last updated: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let uri = item.uri {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "link").symbolRenderingMode(.hierarchical)
-                            Link(destination: uri, label: {
-                                Text(uri.absoluteString)
-                            })
-                            .help("Open in browser")
-                            .underline()
-                            .useDefaultHover({_ in})
-                            .contextMenu {
-                                Button {
-                                    ClipboardHelper.copy(uri.absoluteString)
-                                } label: {
-                                    Text("Copy to clipboard")
-                                }
-                            }
-                            Spacer()
-                        }
-                        Divider()
-                    }
-                    
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "camera.filters").symbolRenderingMode(.hierarchical)
-                        Rectangle()
-                            .frame(width: 15, height: 15)
-                            .background(item.colour_from_stored())
-                        Spacer()
-                    }
-                    .help("Colour: \(item.colour_from_stored().description)")
-                    Divider()
-
-                    Spacer()
-                    HStack(alignment: .top, spacing: 10) {
-//                        // TODO: throws a "serious application error" on load, issue probably in JobDashboard tho
-//                        FancyButtonv2(
-//                            text: "Open",
-//                            icon: "arrow.right.square.fill",
-//                            showLabel: true,
-//                            size: .link,
-//                            type: .clear,
-//                            redirect: AnyView(JobDashboard(defaultSelectedJob: item)),
-//                            pageType: .jobs,
-//                            sidebar: AnyView(JobDashboardSidebar())
-//                        )
-                        FancyButtonv2(
-                            text: nav.session.job != nil ?
-                                (
-                                    nav.session.job == item ? "Current job" : "Overwrite Active Job"
-                                ):
-                                "Set to Active Job",
-                            action: {nav.session.job = item},
-                            icon: "arrow.right.square.fill",
-                            showLabel: true,
-                            size: .link,
-                            type: .clear
-                        )
-                        .disabled(nav.session.job == item)
-                        .help(nav.session.job != nil ? "Current: \(nav.session.job!.jid)" : "Flags this as the current job on other pages and in widgets.")
-                    }
-                }
-            }
-        }
-        
-        struct InspectingRecord: View {
-            public var item: LogRecord
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Record")
-                        Spacer()
-                    }
-                    .help("Type: Record entity")
-                    Divider()
-                    
-                    if let date = item.timestamp {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let message = item.message {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "doc.text.fill").symbolRenderingMode(.hierarchical)
-                            Text("Full message:")
-                            Spacer()
-                        }
-                        Divider()
-                        Text(message)
-                            .padding(.leading, 25)
-                            .contextMenu {
-                                Button("Copy") {
-                                    ClipboardHelper.copy(message)
-                                }
-                            }
-                        .help("Full contents of this message")
-                    }
-                    
-                    Divider()
-                    
-                    Spacer()
-                    HStack(alignment: .top, spacing: 10) {
-                        FancyButtonv2(
-                            text: "Open day",
-                            action: {nav.session.date = item.timestamp ?? Date()},
-                            icon: "arrow.right.square.fill",
-                            showLabel: true,
-                            size: .link,
-                            type: .clear,
-                            redirect: AnyView(Today()),
-                            pageType: .today,
-                            sidebar: AnyView(TodaySidebar())
-                        )
-                    }
-                }
-            }
-        }
-        
-        struct InspectingProject: View {
-            public var item: Project
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Project")
-                        Spacer()
-                    }
-                    .help("Type: Project entity")
-                    Divider()
-                    
-                    if let date = item.created {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    Spacer()
-                    if let company = item.company {
-                        HStack(alignment: .top, spacing: 10) {
-                            FancyButtonv2(
-                                text: "Open project",
-                                icon: "arrow.right.square.fill",
-                                showLabel: true,
-                                size: .link,
-                                type: .clear,
-                                redirect: AnyView(CompanyDashboard(company: company)),
-                                pageType: .companies,
-                                sidebar: AnyView(DefaultCompanySidebar())
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        
-        struct InspectingCompany: View {
-            public var item: Company
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Company")
-                        Spacer()
-                    }
-                    .help("Type: Company entity")
-                    Divider()
-                    
-                    if let date = item.createdDate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    Spacer()
-                    HStack(alignment: .top, spacing: 10) {
-                        FancyButtonv2(
-                            text: "Open company",
-                            icon: "arrow.right.square.fill",
-                            showLabel: true,
-                            size: .link,
-                            type: .clear,
-                            redirect: AnyView(CompanyDashboard(company: item)),
-                            pageType: .companies,
-                            sidebar: AnyView(DefaultCompanySidebar())
-                        )
-                    }
-                }
-            }
-        }
-        
-        struct InspectingPerson: View {
-            public var item: Person
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Person")
-                        Spacer()
-                    }
-                    .help("Type: Person entity")
-                    Divider()
-                    
-                    if let date = item.created {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.lastUpdate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Last update: \(date.description)")
-                        Divider()
-                    }
-                    
-                    Spacer()
-                    if let company = item.company {
-                        HStack(alignment: .top, spacing: 10) {
-                            FancyButtonv2(
-                                text: "Open",
-                                icon: "arrow.right.square.fill",
-                                showLabel: true,
-                                size: .link,
-                                type: .clear,
-                                redirect: AnyView(CompanyDashboard(company: company)),
-                                pageType: .companies,
-                                sidebar: AnyView(DefaultCompanySidebar())
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        
-        struct InspectingNote: View {
-            public var item: Note
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Note")
-                        Spacer()
-                    }
-                    .help("Type: Note entity")
-                    Divider()
-                    
-                    if let date = item.postedDate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.lastUpdate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Last update: \(date.description)")
-                        Divider()
-                    }
-                    
-                    Spacer()
-                    HStack(alignment: .top, spacing: 10) {
-                        FancyButtonv2(
-                            text: "Open",
-                            action: {nav.session.search.cancel()},
-                            icon: "arrow.right.square.fill",
-                            showLabel: true,
-                            size: .link,
-                            type: .clear,
-                            redirect: AnyView(NoteView(note: item, moc: PersistenceController.shared.container.viewContext)),
-                            pageType: .notes,
-                            sidebar: AnyView(NoteViewSidebar(note: item, moc: PersistenceController.shared.container.viewContext))
-                        )
-                    }
-                }
-            }
-        }
-        
-        struct InspectingTask: View {
-            public var item: LogTask
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                VStack(alignment: .leading, spacing: 10) {
-                    HStack(alignment: .top, spacing: 10) {
-                        Image(systemName: "questionmark.square.fill").symbolRenderingMode(.hierarchical)
-                        Text("Type: Task")
-                        Spacer()
-                    }
-                    .help("Type: Task entity")
-                    Divider()
-                    
-                    if let date = item.created {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.plus").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Created: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.lastUpdate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Last update: \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.completedDate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Completed on \(date.description)")
-                        Divider()
-                    }
-                    
-                    if let date = item.cancelledDate {
-                        HStack(alignment: .top, spacing: 10) {
-                            Image(systemName: "calendar.badge.clock").symbolRenderingMode(.hierarchical)
-                            Text(date.description)
-                            Spacer()
-                        }
-                        .help("Cancelled on \(date.description)")
-                        Divider()
-                    }
-                    
-                    Spacer()
-                    if let job = item.owner {
-                        if let project = job.project {
-                            HStack(alignment: .top, spacing: 10) {
-                                FancyButtonv2(
-                                    text: "Open",
-                                    icon: "arrow.right.square.fill",
-                                    showLabel: true,
-                                    size: .link,
-                                    type: .clear,
-                                    redirect: AnyView(TaskDashboardByProject(project: project)),
-                                    pageType: .companies,
-                                    sidebar: AnyView(DefaultCompanySidebar())
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    struct Suggestions: View {
-        @Binding public var searchText: String
-        @Binding public var publishedOnly: Bool
-        @Binding public var showRecords: Bool
-        @Binding public var showNotes: Bool
-        @Binding public var showTasks: Bool
-        @Binding public var showProjects: Bool
-        @Binding public var showJobs: Bool
-        @Binding public var showCompanies: Bool
-        @Binding public var showPeople: Bool
-        public var location: WidgetLocation
-
-        @Environment(\.managedObjectContext) var moc
-        @EnvironmentObject public var nav: Navigation
-
-        var body: some View {
-            VStack {
-                if searchText.count >= 2 {
-                    ScrollView(showsIndicators: false) {
-                        VStack {
-                            HStack {
-                                Text("Hit \"return\" to perform a search and see all results")
-                                Spacer()
-                            }
-                            
-                            // @TODO: reduce this with a loop, each view is basically identical...
-                            if showJobs {SuggestedJobs(searchText: $searchText, publishedOnly: $publishedOnly)}
-                            if showProjects {SuggestedProjects(searchText: $searchText, publishedOnly: $publishedOnly)}
-                            if showNotes {SuggestedNotes(searchText: $searchText, publishedOnly: $publishedOnly)}
-                            if showTasks {SuggestedTasks(searchText: $searchText)}
-                            if showRecords {SuggestedRecords(searchText: $searchText, publishedOnly: $publishedOnly)}
-                            if showCompanies {SuggestedCompanies(searchText: $searchText, publishedOnly: $publishedOnly)}
-                            if showPeople {SuggestedPeople(searchText: $searchText)}
-                        }
-                    }
-                    .padding()
-                }
-            }
-            .background(location == .content ? Theme.rowColour : Color.clear)
-        }
-        
-        struct SuggestedJobs: View {
-            @Binding public var searchText: String
-            @Binding public var publishedOnly: Bool
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<Job>
-            
-            @EnvironmentObject public var nav: Navigation
-            
-            var body: some View {
-                if items.count > 0 {
-                    VStack(alignment: .leading) {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Jobs")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-                        
-                        if showChildren {
-                            VStack(alignment: .leading) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack(alignment: .leading, spacing: 10) {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.jid.string,
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill", 
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                            FancyButtonv2(
-                                                text: item.jid.string,
-                                                action: {nav.session.setJob(item)},
-                                                icon: "arrow.right.square.fill",
-                                                fgColour: .white,
-                                                showLabel: false,
-                                                showIcon: true,
-                                                size: .tinyLink,
-                                                type: .clear
-                                            )
-                                            .help("Set as Active Job")
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>, publishedOnly: Binding<Bool>) {
-                _searchText = searchText
-                _publishedOnly = publishedOnly
-                
-                let req: NSFetchRequest<Job> = Job.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \Job.created, ascending: false),
-                ]
-
-                if publishedOnly.wrappedValue {
-                    req.predicate = NSPredicate(format: "alive = true && jid.string BEGINSWITH %@", _searchText.wrappedValue)
-                } else {
-                    req.predicate = NSPredicate(format: "jid.string BEGINSWITH %@", _searchText.wrappedValue)
-                }
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedProjects: View {
-            @Binding public var searchText: String
-            @Binding public var publishedOnly: Bool
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<Project>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Projects")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.name ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>, publishedOnly: Binding<Bool>) {
-                _searchText = searchText
-                _publishedOnly = publishedOnly
-                
-                let req: NSFetchRequest<Project> = Project.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \Project.name, ascending: true),
-                ]
-                
-                if publishedOnly.wrappedValue {
-                    req.predicate = NSPredicate(
-                        format: "alive = true && (name BEGINSWITH %@ || pid BEGINSWITH %@)",
-                        _searchText.wrappedValue,
-                        _searchText.wrappedValue
-                    )
-                } else {
-                    req.predicate = NSPredicate(
-                        format: "name BEGINSWITH %@ || pid BEGINSWITH %@",
-                        _searchText.wrappedValue,
-                        _searchText.wrappedValue
-                    )
-                }
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedNotes: View {
-            @Binding public var searchText: String
-            @Binding public var publishedOnly: Bool
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<Note>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Notes")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-                        
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.title ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>, publishedOnly: Binding<Bool>) {
-                _searchText = searchText
-                _publishedOnly = publishedOnly
-                
-                let req: NSFetchRequest<Note> = Note.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \Note.postedDate, ascending: false),
-                ]
-                
-                if publishedOnly.wrappedValue {
-                    req.predicate = NSPredicate(
-                        format: "alive = true && (body CONTAINS[cd] %@ || title CONTAINS[cd] %@)",
-                        _searchText.wrappedValue,
-                        _searchText.wrappedValue
-                    )
-                } else {
-                    req.predicate = NSPredicate(
-                        format: "body CONTAINS[cd] %@ || title CONTAINS[cd] %@",
-                        _searchText.wrappedValue,
-                        _searchText.wrappedValue
-                    )
-                }
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedTasks: View {
-            @Binding public var searchText: String
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<LogTask>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Tasks")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.content ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>) {
-                _searchText = searchText
-
-                let req: NSFetchRequest<LogTask> = LogTask.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \LogTask.created, ascending: true),
-                ]
-
-                req.predicate = NSPredicate(
-                    format: "content CONTAINS[cd] %@",
-                    _searchText.wrappedValue
-                )
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedRecords: View {
-            @Binding public var searchText: String
-            @Binding public var publishedOnly: Bool
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<LogRecord>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Records")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.message ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>, publishedOnly: Binding<Bool>) {
-                _searchText = searchText
-                _publishedOnly = publishedOnly
-                
-                let req: NSFetchRequest<LogRecord> = LogRecord.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \LogRecord.timestamp, ascending: false),
-                ]
-                
-                if publishedOnly.wrappedValue {
-                    req.predicate = NSPredicate(
-                        format: "alive = true && message CONTAINS[cd] %@",
-                        _searchText.wrappedValue
-                    )
-                } else {
-                    req.predicate = NSPredicate(
-                        format: "message CONTAINS[cd] %@",
-                        _searchText.wrappedValue
-                    )
-                }
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedCompanies: View {
-            @Binding public var searchText: String
-            @Binding public var publishedOnly: Bool
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<Company>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) Companies")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.name ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>, publishedOnly: Binding<Bool>) {
-                _searchText = searchText
-                _publishedOnly = publishedOnly
-                
-                let req: NSFetchRequest<Company> = Company.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \Company.name, ascending: true),
-                ]
-                
-                if publishedOnly.wrappedValue {
-                    req.predicate = NSPredicate(
-                        format: "alive = true && name CONTAINS[cd] %@",
-                        _searchText.wrappedValue
-                    )
-                } else {
-                    req.predicate = NSPredicate(
-                        format: "name CONTAINS[cd] %@",
-                        _searchText.wrappedValue
-                    )
-                }
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-        
-        struct SuggestedPeople: View {
-            @Binding public var searchText: String
-            @State private var showChildren: Bool = false
-            @State private var hover: Bool = false
-            @FetchRequest private var items: FetchedResults<Person>
-            
-            @EnvironmentObject public var nav: Navigation
-
-            var body: some View {
-                if items.count > 0 {
-                    VStack {
-                        Button {
-                            showChildren.toggle()
-                        } label: {
-                            ZStack {
-                                Theme.base
-                                HStack(spacing: 1) {
-                                    Text("Showing \(items.prefix(5).count)/\(items.count) People")
-                                        .font(Theme.fontSubTitle)
-                                    Spacer()
-                                    Image(systemName: showChildren ? "minus.square.fill" : "plus.square.fill").symbolRenderingMode(.hierarchical)
-                                        .font(.title2)
-                                }
-                                .padding()
-                                .background(hover ? Theme.rowColour : Theme.subHeaderColour)
-                            }
-                        }
-                        .buttonStyle(.plain)
-                        .useDefaultHover({inside in hover = inside})
-
-                        if showChildren {
-                            VStack(alignment: .leading, spacing: 0) {
-                                ForEach(items.prefix(5)) { item in
-                                    VStack {
-                                        Divider()
-                                        HStack {
-                                            FancyButtonv2(
-                                                text: item.name ?? "",
-                                                action: {choose(item)},
-                                                icon: "questionmark.square.fill",
-                                                fgColour: .white,
-                                                showIcon: true,
-                                                size: .link,
-                                                type: .clear
-                                            )
-                                            .help("Inspect")
-                                            Spacer()
-                                        }
-                                    }
-                                    .padding(.bottom, 10)
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    EmptyView()
-                }
-            }
-            
-            init(searchText: Binding<String>) {
-                _searchText = searchText
-                
-                let req: NSFetchRequest<Person> = Person.fetchRequest()
-                req.sortDescriptors = [
-                    NSSortDescriptor(keyPath: \Person.created, ascending: false),
-                ]
-                req.predicate = NSPredicate(
-                    format: "name CONTAINS[cd] %@",
-                    _searchText.wrappedValue
-                )
-                
-                _items = FetchRequest(fetchRequest: req, animation: .easeInOut)
-            }
-        }
-    }
-    
     struct RecordsMatchingString: View {
         @State private var text: String = ""
         @State private var loaded: Bool = false
@@ -1647,48 +512,6 @@ extension FindDashboard {
             ]
             _entities = FetchRequest(fetchRequest: jr, animation: .easeInOut)
         }
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedJobs {
-    private func choose(_ item: Job) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedProjects {
-    private func choose(_ item: Project) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedNotes {
-    private func choose(_ item: Note) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedTasks {
-    private func choose(_ item: LogTask) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedCompanies {
-    private func choose(_ item: Company) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedPeople {
-    private func choose(_ item: Person) -> Void {
-        nav.session.search.inspectingEntity = item
-    }
-}
-
-extension FindDashboard.Suggestions.SuggestedRecords {
-    private func choose(_ item: LogRecord) -> Void {
-        nav.session.search.inspectingEntity = item
     }
 }
